feat: transaksi kas view

+++ app/src/main/java/com/stiproject/kelassti/model/response/transaksi/TransaksiData.kt
@@ -3,7 +3,7 @@ package com.stiproject.kelassti.model.response.transaksi
 data class TransaksiData(
     val id: Int,
     val NIM_mahasiswa: Int,
-    val nominal: Int,
+    val nominal: Double,
     val type: String,
     val tanggal: String,
     val deskripsi: String
